RoomPoints: Fixed bug with sprite location for middle hallways
moveDialog:
If Player was moved to a room during a suggestion, they are able to make a suggestion without moving on their next turn. If they move to a hallway, the suggestion button needs to be disabled.

GameBoard:
Removed unused code

ClientHandler:
Added a broadcast message when a player ends their turn

GameController:
Fixed bug with mark counter skipping the next player in turn order.